ci: switch to per-package coverage thresholds
Replace single 80% total coverage gate with per-package thresholds
that reflect each module's testability profile. Key changes:

- ci.yml: per-package coverage check with declarative threshold table
- ci.yml: exclude cmd/worker and worker/pi from coverage profile
- codecov.yml: lower overall target to 50%, add messaging module at 50%
- Makefile: add `make coverage` target with CI-matching exclusions
- Testing-Strategy.md: update Section 5 with actual package thresholds
- stt.go: fix randomAlphaNum int64 overflow (uint64 cast)
- stt_test.go: comprehensive tests for all STT implementations

+++ internal/messaging/feishu/stt.go
@@ -467,8 +467,9 @@ func randomAlphaNum(n int) string {
 	b := make([]byte, n)
 	for i := range b {
 		// Simple fast generation — file_id is not security-sensitive.
+		// Use uint64 to avoid negative modulo from int64 overflow.
 		now = now*1664525 + 1013904223
-		b[i] = charset[int(now%int64(len(charset)))]
+		b[i] = charset[int(uint64(now)%uint64(len(charset)))]
 	}
 	return string(b)
 }
